Allow tracking task ID to support reporting timeout error details
When a timeout occurs while waiting for a sequence of tasks to finish execution, it is useful to understand which task was executing when the timeout occurred. Using an `Int` as task ID so the `SynchronizedSequenceExecutionHandle` can track it using the efficient lock-free `AtomicInt`, in order to minimize performance cost when tracking is enabled.